third_party/WebKit/LayoutTests/fast/css-grid-layout/grid-item-start-before-get-set-expected.txt

Test that setting and getting grid-column-start and grid-row-start works as expected

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Test getting grid-column-start and grid-row-start set through CSS
PASS getComputedStyle(gridElement, '').getPropertyValue('grid-column-start') is 'auto'
PASS getComputedStyle(gridElement, '').getPropertyValue('grid-column') is 'auto / auto'
PASS getComputedStyle(gridElement, '').getPropertyValue('grid-row-start') is 'auto'
PASS getComputedStyle(gridElement, '').getPropertyValue('grid-row') is 'auto / auto'
PASS getComputedStyle(gridItemWithPositiveInteger, '').getPropertyValue('grid-column') is "10 / auto"
PASS getComputedStyle(gridItemWithPositiveInteger, '').getPropertyValue('grid-column-start') is "10"
PASS getComputedStyle(gridItemWithPositiveInteger, '').getPropertyValue('grid-column-end') is "auto"
PASS getComputedStyle(gridItemWithPositiveInteger, '').getPropertyValue('grid-row') is "15 / auto"
PASS getComputedStyle(gridItemWithPositiveInteger, '').getPropertyValue('grid-row-start') is "15"
PASS getComputedStyle(gridItemWithPositiveInteger, '').getPropertyValue('grid-row-end') is "auto"
PASS getComputedStyle(gridItemWithNegativeInteger, '').getPropertyValue('grid-column') is "-10 / auto"
PASS getComputedStyle(gridItemWithNegativeInteger, '').getPropertyValue('grid-column-start') is "-10"
PASS getComputedStyle(gridItemWithNegativeInteger, '').getPropertyValue('grid-column-end') is "auto"
PASS getComputedStyle(gridItemWithNegativeInteger, '').getPropertyValue('grid-row') is "-15 / auto"
PASS getComputedStyle(gridItemWithNegativeInteger, '').getPropertyValue('grid-row-start') is "-15"
PASS getComputedStyle(gridItemWithNegativeInteger, '').getPropertyValue('grid-row-end') is "auto"
PASS getComputedStyle(gridItemWithBeforeSpan, '').getPropertyValue('grid-column') is "span 2 / auto"
PASS getComputedStyle(gridItemWithBeforeSpan, '').getPropertyValue('grid-column-start') is "span 2"
PASS getComputedStyle(gridItemWithBeforeSpan, '').getPropertyValue('grid-column-end') is "auto"
PASS getComputedStyle(gridItemWithBeforeSpan, '').getPropertyValue('grid-row') is "span 8 / auto"
PASS getComputedStyle(gridItemWithBeforeSpan, '').getPropertyValue('grid-row-start') is "span 8"
PASS getComputedStyle(gridItemWithBeforeSpan, '').getPropertyValue('grid-row-end') is "auto"
PASS getComputedStyle(gridItemWithAfterSpan, '').getPropertyValue('grid-column') is "span 2 / auto"
PASS getComputedStyle(gridItemWithAfterSpan, '').getPropertyValue('grid-column-start') is "span 2"
PASS getComputedStyle(gridItemWithAfterSpan, '').getPropertyValue('grid-column-end') is "auto"
PASS getComputedStyle(gridItemWithAfterSpan, '').getPropertyValue('grid-row') is "span 8 / auto"
PASS getComputedStyle(gridItemWithAfterSpan, '').getPropertyValue('grid-row-start') is "span 8"
PASS getComputedStyle(gridItemWithAfterSpan, '').getPropertyValue('grid-row-end') is "auto"
PASS getComputedStyle(gridItemWithOnlySpan, '').getPropertyValue('grid-column') is "auto / auto"
PASS getComputedStyle(gridItemWithOnlySpan, '').getPropertyValue('grid-column-start') is "auto"
PASS getComputedStyle(gridItemWithOnlySpan, '').getPropertyValue('grid-column-end') is "auto"
PASS getComputedStyle(gridItemWithOnlySpan, '').getPropertyValue('grid-row') is "auto / auto"
PASS getComputedStyle(gridItemWithOnlySpan, '').getPropertyValue('grid-row-start') is "auto"
PASS getComputedStyle(gridItemWithOnlySpan, '').getPropertyValue('grid-row-end') is "auto"
PASS getComputedStyle(gridItemWithAutoElement, '').getPropertyValue('grid-column') is "auto / auto"
PASS getComputedStyle(gridItemWithAutoElement, '').getPropertyValue('grid-column-start') is "auto"
PASS getComputedStyle(gridItemWithAutoElement, '').getPropertyValue('grid-column-end') is "auto"
PASS getComputedStyle(gridItemWithAutoElement, '').getPropertyValue('grid-row') is "auto / auto"
PASS getComputedStyle(gridItemWithAutoElement, '').getPropertyValue('grid-row-start') is "auto"
PASS getComputedStyle(gridItemWithAutoElement, '').getPropertyValue('grid-row-end') is "auto"
PASS getComputedStyle(gridItemWithCustomIdent, '').getPropertyValue('grid-column') is "first / auto"
PASS getComputedStyle(gridItemWithCustomIdent, '').getPropertyValue('grid-column-start') is "first"
PASS getComputedStyle(gridItemWithCustomIdent, '').getPropertyValue('grid-column-end') is "auto"
PASS getComputedStyle(gridItemWithCustomIdent, '').getPropertyValue('grid-row') is "last / auto"
PASS getComputedStyle(gridItemWithCustomIdent, '').getPropertyValue('grid-row-start') is "last"
PASS getComputedStyle(gridItemWithCustomIdent, '').getPropertyValue('grid-row-end') is "auto"
PASS getComputedStyle(gridItemWithNumberCustomIdent, '').getPropertyValue('grid-column') is "2 first / auto"
PASS getComputedStyle(gridItemWithNumberCustomIdent, '').getPropertyValue('grid-column-start') is "2 first"
PASS getComputedStyle(gridItemWithNumberCustomIdent, '').getPropertyValue('grid-column-end') is "auto"
PASS getComputedStyle(gridItemWithNumberCustomIdent, '').getPropertyValue('grid-row') is "3 last / auto"
PASS getComputedStyle(gridItemWithNumberCustomIdent, '').getPropertyValue('grid-row-start') is "3 last"
PASS getComputedStyle(gridItemWithNumberCustomIdent, '').getPropertyValue('grid-row-end') is "auto"
PASS getComputedStyle(gridItemWithSpanCustomIdentElement, '').getPropertyValue('grid-column') is "span 1 first / auto"
PASS getComputedStyle(gridItemWithSpanCustomIdentElement, '').getPropertyValue('grid-column-start') is "span 1 first"
PASS getComputedStyle(gridItemWithSpanCustomIdentElement, '').getPropertyValue('grid-column-end') is "auto"
PASS getComputedStyle(gridItemWithSpanCustomIdentElement, '').getPropertyValue('grid-row') is "span 1 last / auto"
PASS getComputedStyle(gridItemWithSpanCustomIdentElement, '').getPropertyValue('grid-row-start') is "span 1 last"
PASS getComputedStyle(gridItemWithSpanCustomIdentElement, '').getPropertyValue('grid-row-end') is "auto"
PASS getComputedStyle(gridItemWithSpanNumberCustomIdentElement, '').getPropertyValue('grid-column') is "span 2 first / auto"
PASS getComputedStyle(gridItemWithSpanNumberCustomIdentElement, '').getPropertyValue('grid-column-start') is "span 2 first"
PASS getComputedStyle(gridItemWithSpanNumberCustomIdentElement, '').getPropertyValue('grid-column-end') is "auto"
PASS getComputedStyle(gridItemWithSpanNumberCustomIdentElement, '').getPropertyValue('grid-row') is "span 3 last / auto"
PASS getComputedStyle(gridItemWithSpanNumberCustomIdentElement, '').getPropertyValue('grid-row-start') is "span 3 last"
PASS getComputedStyle(gridItemWithSpanNumberCustomIdentElement, '').getPropertyValue('grid-row-end') is "auto"
PASS getComputedStyle(gridItemWithArea, '').getPropertyValue('grid-column') is "firstArea / auto"
PASS getComputedStyle(gridItemWithArea, '').getPropertyValue('grid-column-start') is "firstArea"
PASS getComputedStyle(gridItemWithArea, '').getPropertyValue('grid-column-end') is "auto"
PASS getComputedStyle(gridItemWithArea, '').getPropertyValue('grid-row') is "thirdArea / auto"
PASS getComputedStyle(gridItemWithArea, '').getPropertyValue('grid-row-start') is "thirdArea"
PASS getComputedStyle(gridItemWithArea, '').getPropertyValue('grid-row-end') is "auto"


Test the initial value
PASS getComputedStyle(element, '').getPropertyValue('grid-column-start') is 'auto'
PASS getComputedStyle(element, '').getPropertyValue('grid-column') is 'auto / auto'
PASS getComputedStyle(element, '').getPropertyValue('grid-row-start') is 'auto'
PASS getComputedStyle(element, '').getPropertyValue('grid-row') is 'auto / auto'


Test getting and setting grid-column-start and grid-row-start through JS
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column') is "18 / auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-start') is "18"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-end') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row') is "66 / auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-start') is "66"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-end') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column') is "-55 / auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-start') is "-55"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-end') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row') is "-40 / auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-start') is "-40"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-end') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column') is "nav / auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-start') is "nav"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-end') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row') is "last / auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-start') is "last"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-end') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column') is "span 3 / auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-start') is "span 3"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-end') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row') is "span 20 / auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-start') is "span 20"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-end') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column') is "span 1 nav / auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-start') is "span 1 nav"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-end') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row') is "span 1 last / auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-start') is "span 1 last"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-end') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column') is "auto / auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-start') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-end') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row') is "auto / auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-start') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-end') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column') is "thirdArea / auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-start') is "thirdArea"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-end') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row') is "secondArea / auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-start') is "secondArea"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-end') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column') is "nonExistentArea / auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-start') is "nonExistentArea"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-end') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row') is "secondArea / auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-start') is "secondArea"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-end') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column') is "secondArea / auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-start') is "secondArea"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-end') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row') is "nonExistentArea / auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-start') is "nonExistentArea"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-end') is "auto"


Test setting grid-column-start and grid-row-start to 'inherit' through JS
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column') is "6 / auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-start') is "6"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-end') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row') is "18 / auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-start') is "18"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-end') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column') is "2 / auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-start') is "2"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-end') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row') is "span 1 / auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-start') is "span 1"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-end') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column') is "6 / auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-start') is "6"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-end') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row') is "span 1 / auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-start') is "span 1"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-end') is "auto"


Test setting grid-column-start and grid-row-start to 'initial' through JS
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column') is "1 / span 2"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-start') is "1"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-end') is "span 2"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row') is "span 3 / 5"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-start') is "span 3"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-end') is "5"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column') is "auto / span 2"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-start') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-end') is "span 2"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row') is "span 3 / 5"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-start') is "span 3"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-end') is "5"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column') is "auto / span 2"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-start') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-column-end') is "span 2"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row') is "auto / 5"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-start') is "auto"
PASS getComputedStyle(gridItem, '').getPropertyValue('grid-row-end') is "5"


Test setting grid-column-start and grid-row-start back to 'auto' through JS
PASS getComputedStyle(element, '').getPropertyValue('grid-column-start') is '18'
PASS getComputedStyle(element, '').getPropertyValue('grid-column') is '18 / auto'
PASS getComputedStyle(element, '').getPropertyValue('grid-row-start') is '66'
PASS getComputedStyle(element, '').getPropertyValue('grid-row') is '66 / auto'
PASS getComputedStyle(element, '').getPropertyValue('grid-column-start') is 'auto'
PASS getComputedStyle(element, '').getPropertyValue('grid-column') is 'auto / auto'
PASS getComputedStyle(element, '').getPropertyValue('grid-row-start') is 'auto'
PASS getComputedStyle(element, '').getPropertyValue('grid-row') is 'auto / auto'
PASS successfullyParsed is true

TEST COMPLETE
